requirements.txt
opencv-python
qrcode
Pillow
numpy
imageio
